remove custom serialization for optimization from PoolConfig, VaultState, AmplificationState  (#605)

+++ pkg/vault/contracts/factories/FactoryWidePauseWindow.sol
@@ -20,29 +20,28 @@ contract FactoryWidePauseWindow {
     /// @dev The factory deployer gave a duration that would overflow the Unix timestamp.
     error PoolPauseWindowDurationOverflow();
 
-    // The pause window end time is stored in 32 bits.
-    uint256 private constant _MAX_TIMESTAMP = type(uint32).max;
+    uint32 private constant _MAX_TIMESTAMP = type(uint32).max;
 
-    uint256 private immutable _pauseWindowDuration;
+    uint32 private immutable _pauseWindowDuration;
 
     // Time when the pause window for all created Pools expires.
-    uint256 private immutable _poolsPauseWindowEndTime;
+    uint32 private immutable _poolsPauseWindowEndTime;
 
     constructor(uint256 pauseWindowDuration) {
         if (block.timestamp + pauseWindowDuration > _MAX_TIMESTAMP) {
             revert PoolPauseWindowDurationOverflow();
         }
 
-        _pauseWindowDuration = pauseWindowDuration;
+        _pauseWindowDuration = uint32(pauseWindowDuration);
 
-        _poolsPauseWindowEndTime = block.timestamp + pauseWindowDuration;
+        _poolsPauseWindowEndTime = uint32(block.timestamp) + uint32(pauseWindowDuration);
     }
 
     /**
      * @notice Return the pause window duration. This is the time pools will be pausable after factory deployment.
      * @return The duration in seconds
      */
-    function getPauseWindowDuration() external view returns (uint256) {
+    function getPauseWindowDuration() external view returns (uint32) {
         return _pauseWindowDuration;
     }
 
@@ -50,7 +49,7 @@ contract FactoryWidePauseWindow {
      * @notice Returns the original factory pauseWindowEndTime, regardless of the current time.
      * @return The end time as a timestamp
      */
-    function getOriginalPauseWindowEndTime() external view returns (uint256) {
+    function getOriginalPauseWindowEndTime() external view returns (uint32) {
         return _poolsPauseWindowEndTime;
     }
 
@@ -62,7 +61,7 @@ contract FactoryWidePauseWindow {
      *
      * @return The resolved pause window end time (0 indicating it's no longer pausable)
      */
-    function getNewPoolPauseWindowEndTime() public view returns (uint256) {
-        return block.timestamp < _poolsPauseWindowEndTime ? _poolsPauseWindowEndTime : 0;
+    function getNewPoolPauseWindowEndTime() public view returns (uint32) {
+        return uint32(block.timestamp) < _poolsPauseWindowEndTime ? _poolsPauseWindowEndTime : 0;
     }
 }
